Added support for complex Zeta

+++ src/zeta.cpp
@@ -13,22 +13,22 @@
 #include <complex>
 
 
-double S(double s, double a, int N) {
-  double sum = 0.;
+cdouble S(double s, cdouble a, int N) {
+  cdouble sum = 0.;
 
   for (int k = 0; k <= N - 1; k += 1) {
-    sum += pow(a + k, -s);
+    sum += pow(a + static_cast<cdouble>(k), -s);
   }
 
   return sum;
 }
 
-double I(double s, double a, int N) {
-  return pow(a + N, 1. - s) / (s - 1.);
+cdouble I(double s, cdouble a, int N) {
+  return pow(a + static_cast<cdouble>(N), 1. - s) / (s - 1.);
 }
 
-double T(double s, double a, int N, int M) {
-  const double factor = pow(a + N, -s);
+cdouble T(double s, cdouble a, int N, int M) {
+  const cdouble factor = pow(a + static_cast<cdouble>(N), -s);
 
   if (M > B_2n_fact_size) {
     #ifdef DEBUG
@@ -42,17 +42,16 @@ double T(double s, double a, int N, int M) {
     M = B_2n_fact_size;
   }
 
-
-  double sum = 0.;
+  cdouble sum = 0.;
 
   for (int k = 1; k <= M; k += 1) {
-    sum += B_2n_fact[k] * gsl_sf_gamma(s + 2. * k - 1.) / gsl_sf_gamma(s) / pow(a + N, 2. * k - 1.);
+    sum += B_2n_fact[k] * gsl_sf_gamma(s + 2. * k - 1.) / gsl_sf_gamma(s) / pow(a + static_cast<cdouble>(N), 2. * k - 1.);
   }
 
   return factor * (0.5 + sum);
 }
 
-double hurwitz_zeta(double s, double a, int N) {
+cdouble hurwitz_zeta(double s, double a, int N) {
   if (N > B_2n_fact_size) {
     #ifdef DEBUG
       printf("N = %d > B_2n_fact_size = %d. Using N = B_2n_fact_size\n");
